feat : filter -> filterBar로 부모 컴포넌트 추가 후 파일 이동
